\name{git_amped_R-package}
\alias{git_amped_R-package}
\alias{git_amped_R}
\docType{package}
\title{
\packageTitle{git_amped_R}
}
\description{
\packageDescription{git_amped_R}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{git_amped_R}
\packageIndices{git_amped_R}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{git_amped_R}

Maintainer: \packageMaintainer{git_amped_R}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
