\name{git.amped.R-package}
\alias{git.amped.R-package}
\alias{git.amped.R}
\docType{package}
\title{
\packageTitle{git.amped.R}
}
\description{
\packageDescription{git.amped.R}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{git.amped.R}
\packageIndices{git.amped.R}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{git.amped.R}

Maintainer: \packageMaintainer{git.amped.R}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
